refactor: move impl snapshot3 from stream for impl_/stream
